moving AppenderContext to watchdog api
