+++ cdap-watchdog/src/main/java/co/cask/cdap/logging/plugins/FixedWindowRollingPolicy.java
@@ -20,8 +20,8 @@ import ch.qos.logback.core.CoreConstants;
 import ch.qos.logback.core.rolling.RolloverFailure;
 import ch.qos.logback.core.rolling.helper.FileNamePattern;
 import ch.qos.logback.core.rolling.helper.IntegerTokenConverter;
+import co.cask.cdap.api.logging.AppenderContext;
 import co.cask.cdap.common.conf.Constants;
-import co.cask.cdap.logging.framework.AppenderContext;
 import org.apache.twill.filesystem.Location;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
