+++ cdap-watchdog/src/main/java/co/cask/cdap/logging/plugins/RollingLocationLogAppender.java
@@ -23,8 +23,8 @@ import ch.qos.logback.core.rolling.RollingPolicy;
 import ch.qos.logback.core.rolling.RolloverFailure;
 import ch.qos.logback.core.rolling.TriggeringPolicy;
 import ch.qos.logback.core.spi.FilterReply;
+import co.cask.cdap.api.logging.AppenderContext;
 import co.cask.cdap.common.conf.Constants;
-import co.cask.cdap.logging.framework.AppenderContext;
 import co.cask.cdap.proto.id.NamespaceId;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
